Remove filtering from OSS entirely; make it Pro-only end-to-end
Filtering (search) is a Pro-only feature, but its wiring still lived in
OSS: the FilteringHelper, the current_filter/current_filter_field params
readers, and dead filter_box/filtering? branches in OSS views.

- Strip filtering from the OSS views (jobs, cluster brokers, consumers,
  routing index + consumer group); give Pro its own cluster/replication
  view (the one shared view Pro rendered) so the OSS copy is filter-free.
- Move FilteringHelper to Pro (Web::Pro::Ui::Helpers), included by the
  Pro app instead of the shared base.
- Move current_filter/current_filter_field to a Pro Params subclass the
  Pro app builds, dropping them from the OSS params.

Also fix a pre-existing flaky: the cluster filtering tests stubbed
ClusterInfo with topics lacking partition_count, so the link crawler
500'd on /topics when it hit that stub first (order-dependent).
