+++ lib/karafka/web/pro/ui/helpers/filtering_helper.rb
@@ -1,0 +1,235 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Web
+    module Pro
+      module Ui
+        module Helpers
+          # Helpers for rendering the keyword filtering (search) box on data tables. Companion to the
+          # [[SortingHelper]]: sorting reorders a listing, filtering narrows it down.
+          module FilteringHelper
+            # @return [Boolean] true if a filtering keyword is currently active. Filtering (search) is
+            #   a Pro-only feature, so this is always false in OSS.
+            def filtering?
+              return false unless ::Karafka.pro?
+
+              !params.current_filter.empty?
+            end
+
+            # Renders a filtering form for the current listing.
+            #
+            # The form submits via GET to the current path, preserving every other query parameter
+            # (most importantly the current sort) as hidden fields while resetting pagination, so
+            # filtering always starts from the first page.
+            #
+            # It renders as a full-width search field with a "Search" submit button and a "Reset"
+            # button (always present, but disabled when there is nothing to reset, so the layout does
+            # not shift). When the controller exposes a `{ field => label }` map of filterable fields
+            # (via `filter`), a field selector is fused into the left of the field so the user can
+            # pick which attribute to filter on.
+            #
+            # The placeholder is intentionally a single generic default across every listing: a field
+            # selector always accompanies the input and already tells the user which attribute is
+            # being filtered, so a per-view placeholder would be redundant (and would look stale until
+            # the form is submitted, since it is server-rendered and does not follow the select live).
+            #
+            # @param labels [Hash] per-view overrides for the field selector labels, e.g.
+            #   `labels: { id: "Task ID" }`. Anything not overridden falls back to the shared
+            #   {FILTER_NAMES} map (and then to a humanized attribute name).
+            # @return [String] html of the filtering form
+            def filter_box(labels: {})
+              # Filtering (search) is a Pro-only feature, so we render nothing in OSS.
+              return "" unless ::Karafka.pro?
+
+              hidden = preserved_filter_params.map do |key, value|
+                "<input type=\"hidden\" name=\"#{h(key)}\" value=\"#{h(value)}\">"
+              end.join
+
+              # The reset button is always rendered so the layout (and the search button position)
+              # stays put whether or not a filter is active. It is only disabled when there is
+              # nothing to reset.
+              reset =
+                if filtering?
+                  %(<a class="btn btn-outline font-normal" href="#{h(filter_clear_path)}">Reset</a>)
+                else
+                  %(<a class="btn btn-outline font-normal btn-disabled" aria-disabled="true">Reset</a>)
+                end
+
+              # A field is either a plain attribute name or a key-alias descriptor; the selector only
+              # cares about its name
+              fields = Array(@filterable_fields).map { |field| filter_field_name(field) }
+              # We render the field selector whenever the controller exposes any filterable field, so
+              # every search looks consistent (even single-field listings show a one-option select).
+              fielded = fields.any?
+              value_name = fielded ? "filter[value]" : "filter"
+              selected = selected_filter_field(fields) if fielded
+
+              input = <<~HTML
+                <input
+                  type="text"
+                  name="#{value_name}"
+                  value="#{h(params.current_filter)}"
+                  placeholder="Filter..."
+                  class="input w-full#{fielded ? " join-item" : " grow"}"
+                  autocomplete="off"
+                >
+              HTML
+
+              control =
+                if fielded
+                  %(<div class="join grow">#{filter_field_selector(fields, selected, labels)}#{input}</div>)
+                else
+                  input
+                end
+
+              <<~HTML
+                <form method="get" action="#{h(request.path)}" class="filter-form flex gap-2 mb-6">
+                  #{hidden}
+                  #{control}
+                  <button type="submit" class="btn btn-primary font-normal">Search</button>
+                  #{reset}
+                </form>
+              HTML
+            end
+
+            # Non-destructively narrows a topics collection (routing or a consumer subscription) down
+            # to the ones matching the current filter.
+            #
+            # It is field aware: when the `topic` field is selected only the topic names are matched;
+            # when `consumer_group`/`subscription_group` is selected the whole collection is kept or
+            # dropped based on that group's name. With no explicit field (plain keyword) a topic is
+            # kept when its own name matches or when any of the provided group labels match.
+            #
+            # This is used for views that render the live `Karafka::App.routes` (which we must never
+            # mutate, unlike the per-request structures the [[Filter]] engine prunes in place) as well
+            # as the per-process subscriptions view, so it always returns a new array and leaves the
+            # source untouched.
+            #
+            # @param topics [Enumerable] topics of a subscription/consumer group
+            # @param consumer_group [String, nil] the consumer group name shown as the section header
+            # @param subscription_group [String, nil] the subscription group name (subscriptions view)
+            # @return [Array] all topics when no filtering is active, otherwise only the matching ones
+            def visible_topics(topics, consumer_group: nil, subscription_group: nil)
+              return topics.to_a unless filtering?
+
+              keyword = params.current_filter.downcase
+              by_name = -> { topics.select { |topic| topic.name.to_s.downcase.include?(keyword) } }
+              label_match = ->(label) { label.to_s.downcase.include?(keyword) }
+
+              case params.current_filter_field
+              when "consumer_group"
+                label_match.call(consumer_group) ? topics.to_a : []
+              when "subscription_group"
+                label_match.call(subscription_group) ? topics.to_a : []
+              when "topic"
+                by_name.call
+              else
+                # Plain keyword: match the topic name or any of the group labels
+                [consumer_group, subscription_group].any?(&label_match) ? topics.to_a : by_name.call
+              end
+            end
+
+            # Human friendly labels for the attributes we allow filtering on. Attributes not listed
+            # here fall back to a humanized version of their name (see {#filter_field_label}), so
+            # controllers only need to declare bare attribute names in `filterable_attributes`.
+            FILTER_NAMES = {
+              id: "Process ID",
+              subscribed_topics: "Assigned topics",
+              tags: "Tags",
+              topic: "Topic",
+              topic_name: "Topic",
+              consumer_group: "Consumer group",
+              subscription_group: "Subscription group",
+              consumer: "Consumer",
+              type: "Type",
+              name: "Name",
+              value: "Value",
+              cron: "Cron",
+              broker_name: "Broker"
+            }.freeze
+
+            private_constant :FILTER_NAMES
+
+            private
+
+            # @param field [String, Symbol, #name] a filterable field: a plain attribute name or a
+            #   key-alias descriptor
+            # @return [String] the field's name
+            def filter_field_name(field)
+              field.respond_to?(:path) ? field.name.to_s : field.to_s
+            end
+
+            # @param attribute [String, Symbol] filterable attribute name
+            # @param overrides [Hash] per-view label overrides (keyed by attribute)
+            # @return [String] human friendly label for the attribute: a per-view override if given,
+            #   otherwise the shared {FILTER_NAMES} entry, otherwise a humanized attribute name
+            def filter_field_label(attribute, overrides = {})
+              key = attribute.to_sym
+
+              overrides.transform_keys(&:to_sym).fetch(key) do
+                FILTER_NAMES.fetch(key) do
+                  attribute.to_s.tr("_", " ").split.map(&:capitalize).join(" ")
+                end
+              end
+            end
+
+            # @param fields [Array<String, Symbol>] filterable attributes
+            # @return [String] the currently selected filter field, defaulting to the first one when
+            #   nothing (valid) is selected
+            def selected_filter_field(fields)
+              selected = params.current_filter_field
+
+              fields.map(&:to_s).include?(selected) ? selected : fields.first.to_s
+            end
+
+            # Renders the field selector fused into the left of the filtering field.
+            #
+            # @param fields [Array<String, Symbol>] filterable attributes
+            # @param selected [String] the currently selected field
+            # @param overrides [Hash] per-view label overrides (keyed by attribute)
+            # @return [String] html of the select element
+            def filter_field_selector(fields, selected, overrides = {})
+              options = fields.map do |field|
+                key = field.to_s
+                chosen = (key == selected) ? " selected" : ""
+                "<option value=\"#{h(key)}\"#{chosen}>#{h(filter_field_label(field, overrides))}</option>"
+              end.join
+
+              <<~HTML
+                <select name="filter[field]" class="select join-item">
+                  #{options}
+                </select>
+              HTML
+            end
+
+            # Builds the path used by the "clear" control: the current path with all filtering and
+            # pagination parameters removed.
+            #
+            # @return [String] path without any filter/page query parameters
+            def filter_clear_path
+              query = URI.encode_www_form(preserved_filter_params)
+
+              query.empty? ? request.path : "#{request.path}?#{query}"
+            end
+
+            # @return [Hash] current request query parameters that we want to carry over when
+            #   submitting the filtering form (everything except the filtering parameters themselves
+            #   and the pagination, which we intentionally reset)
+            def preserved_filter_params
+              flatten_params("", request.params).reject do |key, _value|
+                key == "filter" || key.start_with?("filter[") || key == "page"
+              end
+            end
+
+            # Escapes a value for safe inclusion in an HTML attribute
+            # @param value [Object] value to escape
+            # @return [String] escaped value
+            def h(value)
+              Rack::Utils.escape_html(value.to_s)
+            end
+          end
+        end
+      end
+    end
+  end
+end
