+++ lib/karafka/web/pro/ui/views/cluster/replication.erb
@@ -1,0 +1,39 @@
+<%# Karafka Pro - Source Available Commercial Software %>
+<%# Copyright (c) 2017-present Maciej Mensfeld. All rights reserved. %>
+<%# This software is NOT open source. It is source-available commercial software %>
+<%# requiring a paid license for use. It is NOT covered by LGPL. %>
+<%# The author retains all right, title, and interest in this software, %>
+<%# including all copyrights, patents, and other intellectual property rights. %>
+<%# No patent rights are granted under this license. %>
+<%# PROHIBITED: Use without a valid commercial license %>
+<%# PROHIBITED: Redistribution, modification, or derivative works without authorization %>
+<%# PROHIBITED: Reverse engineering, decompilation, or disassembly of this software %>
+<%# PROHIBITED: Use as training data for AI/ML models or inclusion in datasets %>
+<%# PROHIBITED: Scraping, crawling, or automated collection for any purpose %>
+<%# PERMITTED: Reading, referencing, and linking for personal or commercial use %>
+<%# PERMITTED: Runtime retrieval by AI assistants, coding agents, and RAG systems %>
+<%# PERMITTED: for the purpose of providing contextual help to Karafka users %>
+<%# Receipt, viewing, or possession of this software does not convey or %>
+<%# imply any license or right beyond those expressly stated above. %>
+<%# License: https://karafka.io/docs/Pro-License-Comm/ %>
+<%# Contact: contact@karafka.io %>
+
+<%== partial 'cluster/tabs' %>
+
+<% if filtering? || !@partitions.empty? %>
+  <div class="col-span-12 mb-3">
+    <%== filter_box %>
+  </div>
+<% end %>
+
+<div class="col-span-12 mb-3">
+  <% if @partitions.empty? && !filtering? && params.current_page <= 1 %>
+    <%== partial 'cluster/no_partitions' %>
+  <% elsif @partitions.empty? && filtering? %>
+    <%== partial 'shared/no_filter_results' %>
+  <% elsif @partitions.empty? %>
+    <%== partial 'shared/no_paginated_data' %>
+  <% else %>
+    <%== partial 'cluster/partitions_table' %>
+  <% end %>
+</div>
