+++ lib/karafka/web/ui/views/cluster/brokers.erb
@@ -2,16 +2,4 @@
 
 <%== partial 'cluster/tabs' %>
 
-<% if Karafka.pro? && (filtering? || !@brokers.empty?) %>
-  <div class="col-span-12 mb-3">
-    <%== filter_box %>
-  </div>
-<% end %>
-
-<% if @brokers.empty? && filtering? %>
-  <div class="col-span-12">
-    <%== partial 'shared/no_filter_results' %>
-  </div>
-<% else %>
-  <%== partial 'cluster/brokers_table' %>
-<% end %>
+<%== partial 'cluster/brokers_table' %>
