+++ lib/karafka/web/ui/views/consumers/index.erb
@@ -4,14 +4,8 @@
 <%== partial 'consumers/tabs' %>
 
 <div class="col-span-12">
-  <% if !@processes.empty? || filtering? %>
-    <%== filter_box %>
-  <% end %>
-
-  <% if @processes.empty? && !filtering? && params.current_page <= 1 %>
+  <% if @processes.empty? && params.current_page <= 1 %>
     <%== partial 'consumers/no_consumers' %>
-  <% elsif @processes.empty? && filtering? %>
-    <%== partial 'shared/no_filter_results' %>
   <% elsif @processes.empty? %>
     <%== partial 'shared/no_paginated_data' %>
   <% else %>
