+++ lib/karafka/web/ui/views/jobs/running.erb
@@ -3,14 +3,8 @@
 <%== partial 'jobs/tabs' %>
 
 <div class="col-span-12">
-  <% if !@jobs.empty? || filtering? %>
-    <%== filter_box %>
-  <% end %>
-
-  <% if @jobs.empty? && !filtering? && params.current_page <= 1 %>
+  <% if @jobs.empty? && params.current_page <= 1 %>
     <%== partial 'jobs/no_jobs', locals: { type: 'running' } %>
-  <% elsif @jobs.empty? && filtering? %>
-    <%== partial 'shared/no_filter_results' %>
   <% elsif @jobs.empty? %>
     <%== partial 'shared/no_paginated_data' %>
   <% else %>
