+++ lib/karafka/web/ui/views/routing/_consumer_group.erb
@@ -1,41 +1,39 @@
 <%
-  visible_pairs = consumer_group.subscription_groups.flat_map do |subscription_group|
-    visible_topics(subscription_group.topics, consumer_group: consumer_group.id, subscription_group: subscription_group.id).map do |topic|
+  pairs = consumer_group.subscription_groups.flat_map do |subscription_group|
+    subscription_group.topics.map do |topic|
       [subscription_group, topic]
     end
   end
 %>
 
-<% unless filtering? && visible_pairs.empty? %>
-  <div class="col-span-12 mb-3">
-    <h2 class="h2">
-      <%= consumer_group.id %>
-    </h2>
+<div class="col-span-12 mb-3">
+  <h2 class="h2">
+    <%= consumer_group.id %>
+  </h2>
 
-    <div class="data-table-wrapper">
-      <table class="data-table">
-        <thead>
-          <tr>
-            <th>Subscription group</th>
-            <th><%== sort_link('Topic', :name) %></th>
-            <th><%== sort_link(:active?) %></th>
-            <th class="options">Options</th>
-          </tr>
-        </thead>
-        <tbody>
-          <% visible_pairs.each do |subscription_group, topic| %>
-            <%==
-              partial(
-                'routing/topic',
-                locals: {
-                  subscription_group: subscription_group,
-                  topic: topic
-                }
-              )
-            %>
-          <% end %>
-        </tbody>
-      </table>
-    </div>
+  <div class="data-table-wrapper">
+    <table class="data-table">
+      <thead>
+        <tr>
+          <th>Subscription group</th>
+          <th><%== sort_link('Topic', :name) %></th>
+          <th><%== sort_link(:active?) %></th>
+          <th class="options">Options</th>
+        </tr>
+      </thead>
+      <tbody>
+        <% pairs.each do |subscription_group, topic| %>
+          <%==
+            partial(
+              'routing/topic',
+              locals: {
+                subscription_group: subscription_group,
+                topic: topic
+              }
+            )
+          %>
+        <% end %>
+      </tbody>
+    </table>
   </div>
-<% end %>
+</div>
