+++ lib/karafka/web/ui/views/routing/index.erb
@@ -1,11 +1,5 @@
 <% view_title 'Routing Details' %>
 
-<% if Karafka.pro? && (filtering? || @routes.any?) %>
-  <div class="col-span-12 mb-3">
-    <%== filter_box %>
-  </div>
-<% end %>
-
 <%==
   each_partial(
     @routes,
